Moved `async_cache` decorator from `Doc` cog file to `utils/cache.py`
